Project 1
Project 1 has three parts to it, the first is reducing the number of gray levels in an image by integer powers of 2.  The second part involved in zooming and shrinking images by pixel replication.  The third part is to develop a 2-D FFT program.  Each part was coded from scratch and did not use built in MATLAB functions, except for the third part where the only function used was fft2 and ifft2.